Moved the encode_*_type functions from assembler.rs to encoder.rs
