+++ README.md
@@ -1,0 +1,2 @@
+# RISC-V Assembler
+An assembler for the RISC-V ISA with support for the RV32I base instruction set, written in Rust.
