+++ src/assembler/mod.rs
@@ -1,3 +1,4 @@
 pub mod assembler;
 mod error;
 mod instructions;
+pub mod encoder;
